Navbar search responsive
Made the Navbar search responsive.  Removed the "close glyph" as well as
the close button.  There was going to be a fair bit of styling needed
for it, and I think it wasn't really needed.  There is some !important
CSS happening to get the typeahead suggestions to show up in the right
place on small viewports, basically needed to override some styling to
make the position of the suggestions fised.