Fix sorting of dates in Firefox
The dates in the tables were in a format that IE and Chrome sorted
correctly, but Firefox did sort sort them correctly.  This is because of
how different browsers interpret date strings. see here:
http://dygraphs.com/date-formats.html

Now we are using DataTables data-sort functionality to provide a sort
value, which is a unix timestamp.